HandsOnProject/Timesheet/timesheet-agent/back-end/requirements.txt
numpy==1.26.4
flask==3.0.3
python-dotenv==1.0.1
sqlalchemy==2.0.31
pyodbc==5.1.0
spacy==3.7.5
fuzzywuzzy==0.18.0
python-dateutil==2.9.0
langchain==0.2.16
langchain-ollama==0.1.3
langchain-chroma==0.1.2
chromadb==0.5.5
